+++ src/Soundheim/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("Custom Audio Output Device")]
 [assembly: AssemblyProduct("Custom Audio Output Device")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion("1.0.1.0")]
+[assembly: AssemblyFileVersion("1.0.1.0")]